Add files via upload
This is the code so far. We have still to fix how DNA works. Everything else we're still working on should be commented out. What we really would like help with is the carnivore(predator) class. We can make them follow food, but not other creatures. Feedback here would be greatly appreciated. 
Carnivore has been commented out, so if you need, you're gonna have to remove the //